feat(gcore): P0 落地——Position 補 z、move_by 移動收口
- position.h：補 int z（x/y/z 全 int），serialize 存三欄；
  Zone::layers 鍵 int16_t→int 同步對齊（zone.h）。
- movement.h：新增 move_by 作位置變更唯一入口（將來 tile flag
  檢查、空間索引維護掛此）；movement 改吃 Zone&，簽章即
  ZoneSystem 可直接註冊（zone_manager.h 範例同步）。
- Task 1（存檔 magic/version）經使用者裁定放棄；存檔格式
  hard break，舊 .bin 作廢。
- CODE_MAP/CODE_TOUR/tome4 報告引用點同步；stale 名稱
  Zone::map→layers 兩處修正。
- 驗證：medp/medp_static 建置綠＋movement.h 臨時 TU 語法編譯過；
  測試套件仍斷裂未跑。

+++ projects/medp/src/gcore/zone/zone.h
@@ -33,5 +33,5 @@ struct Zone {
     // 直接掛在 Zone 上，不走 ECS——地圖是 zone 的固有結構，不是某個 entity 的屬性。
     // 代價：它不在 registry 裡，所以 registry_io 的 snapshot 不會碰它，
     // 存檔由 serialize/zone_io.h 分兩塊處理。
-    std::map<int16_t, tdarray<Tile>> layers;
+    std::map<int, tdarray<Tile>> layers;
 };
